move options module so can load outside bin package
and fix basic moose test
