+++ src/mlde_data/bin/etl.py
@@ -2,7 +2,7 @@ import logging
 import os
 from pathlib import Path
 from mlde_utils import RAW_MOOSE_VARIABLES_PATH
-from mlde_data.bin.options import DomainOption, CollectionOption
+from mlde_data.options import DomainOption, CollectionOption
 from mlde_data.bin.moose import extract, convert, clean
 from mlde_data.bin.variable import create as create_variable
 from mlde_data.variable import load_config
